:art: Reorder transaction functions
