+++ src/index.js
@@ -24,9 +24,11 @@ import crypto from './crypto';
 import dapp from './transactions/dapp';
 import delegate from './transactions/delegate';
 import multisignature from './transactions/multisignature';
+import multisignatureTransaction from './transactions/multisignatureTransaction';
 import signature from './transactions/signature';
 import transaction from './transactions/transaction';
-import transfer from './transactions/transfer';
+import inTransfer from './transactions/inTransfer';
+import outTransfer from './transactions/outTransfer';
 import vote from './transactions/vote';
 import api from './api/liskApi';
 import slots from './time/slots';
@@ -43,9 +45,11 @@ const lisk = {
 	dapp,
 	delegate,
 	multisignature,
+	multisignatureTransaction,
 	signature,
 	transaction,
-	transfer,
+	inTransfer,
+	outTransfer,
 	vote,
 	api,
 	slots,
