+++ test/transactions/delegate.js
@@ -18,133 +18,131 @@ import slots from '../../src/time/slots';
 
 afterEach(() => sandbox.restore());
 
-describe('delegate module', () => {
-	describe('#createDelegate', () => {
-		const secret = 'secret';
-		const secondSecret = 'second secret';
-		const publicKey = '5d036a858ce89f844491762eb89e2bfbd50a4a0a0da658e4b2628b25b117ae09';
-		const secondPublicKey = '0401c8ac9f29ded9e1e4d5b6b43051cb25b22f27c7b7b35092161e851946f82f';
-		const username = 'test_delegate_1@\\';
-		const fee = 25e8;
-		const timeWithOffset = 38350076;
+describe('#createDelegate', () => {
+	const secret = 'secret';
+	const secondSecret = 'second secret';
+	const publicKey = '5d036a858ce89f844491762eb89e2bfbd50a4a0a0da658e4b2628b25b117ae09';
+	const secondPublicKey = '0401c8ac9f29ded9e1e4d5b6b43051cb25b22f27c7b7b35092161e851946f82f';
+	const username = 'test_delegate_1@\\';
+	const fee = 25e8;
+	const timeWithOffset = 38350076;
 
-		let getTimeWithOffsetStub;
-		let delegateTransaction;
+	let getTimeWithOffsetStub;
+	let delegateTransaction;
 
+	beforeEach(() => {
+		getTimeWithOffsetStub = sandbox.stub(slots, 'getTimeWithOffset').returns(timeWithOffset);
+	});
+
+	describe('without second secret', () => {
 		beforeEach(() => {
-			getTimeWithOffsetStub = sandbox.stub(slots, 'getTimeWithOffset').returns(timeWithOffset);
+			delegateTransaction = createDelegate(secret, username);
 		});
 
-		describe('without second secret', () => {
-			beforeEach(() => {
-				delegateTransaction = createDelegate(secret, username);
+		it('should create a delegate transaction', () => {
+			(delegateTransaction).should.be.ok();
+		});
+
+		it('should use slots.getTimeWithOffset to calculate the timestamp', () => {
+			(getTimeWithOffsetStub.calledWithExactly(undefined)).should.be.true();
+		});
+
+		it('should use slots.getTimeWithOffset with an offset of -10 seconds to calculate the timestamp', () => {
+			const offset = -10;
+			createDelegate(secret, username, null, offset);
+
+			(getTimeWithOffsetStub.calledWithExactly(offset)).should.be.true();
+		});
+
+		describe('returned delegate transaction', () => {
+			it('should be an object', () => {
+				(delegateTransaction).should.be.type('object');
 			});
 
-			it('should create a delegate transaction', () => {
-				(delegateTransaction).should.be.ok();
+			it('should have an id string', () => {
+				(delegateTransaction).should.have.property('id').and.be.type('string');
 			});
 
-			it('should use slots.getTimeWithOffset to calculate the timestamp', () => {
-				(getTimeWithOffsetStub.calledWithExactly(undefined)).should.be.true();
+			it('should have type number equal to 2', () => {
+				(delegateTransaction).should.have.property('type').and.be.type('number').and.equal(2);
 			});
 
-			it('should use slots.getTimeWithOffset with an offset of -10 seconds to calculate the timestamp', () => {
-				const offset = -10;
-				createDelegate(secret, username, null, offset);
-
-				(getTimeWithOffsetStub.calledWithExactly(offset)).should.be.true();
+			it('should have amount number equal to 0', () => {
+				(delegateTransaction).should.have.property('amount').and.be.type('number').and.equal(0);
 			});
 
-			describe('returned delegate transaction', () => {
+			it('should have fee number equal to 25 LSK', () => {
+				(delegateTransaction).should.have.property('fee').and.be.type('number').and.equal(fee);
+			});
+
+			it('should have recipientId equal to null', () => {
+				(delegateTransaction).should.have.property('recipientId').and.be.null();
+			});
+
+			it('should have senderPublicKey hex string equal to sender public key', () => {
+				(delegateTransaction).should.have.property('senderPublicKey').and.be.hexString().and.equal(publicKey);
+			});
+
+			it('should have timestamp number equal to result of slots.getTimeWithOffset', () => {
+				(delegateTransaction).should.have.property('timestamp').and.be.type('number').and.equal(timeWithOffset);
+			});
+
+			it('should have signature hex string', () => {
+				(delegateTransaction).should.have.property('signature').and.be.hexString();
+			});
+
+			it('should be signed correctly', () => {
+				const result = cryptoModule.verifyTransaction(delegateTransaction);
+				(result).should.be.ok();
+			});
+
+			it('should not be signed correctly if modified', () => {
+				delegateTransaction.amount = 100;
+				const result = cryptoModule.verifyTransaction(delegateTransaction);
+				(result).should.be.not.ok();
+			});
+
+			it('should have asset', () => {
+				(delegateTransaction).should.have.property('asset').and.not.be.empty();
+			});
+
+			describe('delegate asset', () => {
 				it('should be an object', () => {
-					(delegateTransaction).should.be.type('object');
+					(delegateTransaction.asset).should.have.property('delegate').and.be.type('object');
 				});
 
-				it('should have an id string', () => {
-					(delegateTransaction).should.have.property('id').and.be.type('string');
-				});
-
-				it('should have type number equal to 2', () => {
-					(delegateTransaction).should.have.property('type').and.be.type('number').and.equal(2);
-				});
-
-				it('should have amount number equal to 0', () => {
-					(delegateTransaction).should.have.property('amount').and.be.type('number').and.equal(0);
-				});
-
-				it('should have fee number equal to 25 LSK', () => {
-					(delegateTransaction).should.have.property('fee').and.be.type('number').and.equal(fee);
-				});
-
-				it('should have recipientId equal to null', () => {
-					(delegateTransaction).should.have.property('recipientId').and.be.null();
-				});
-
-				it('should have senderPublicKey hex string equal to sender public key', () => {
-					(delegateTransaction).should.have.property('senderPublicKey').and.be.hexString().and.equal(publicKey);
-				});
-
-				it('should have timestamp number equal to result of slots.getTimeWithOffset', () => {
-					(delegateTransaction).should.have.property('timestamp').and.be.type('number').and.equal(timeWithOffset);
-				});
-
-				it('should have signature hex string', () => {
-					(delegateTransaction).should.have.property('signature').and.be.hexString();
-				});
-
-				it('should be signed correctly', () => {
-					const result = cryptoModule.verifyTransaction(delegateTransaction);
-					(result).should.be.ok();
-				});
-
-				it('should not be signed correctly if modified', () => {
-					delegateTransaction.amount = 100;
-					const result = cryptoModule.verifyTransaction(delegateTransaction);
-					(result).should.be.not.ok();
-				});
-
-				it('should have asset', () => {
-					(delegateTransaction).should.have.property('asset').and.not.be.empty();
-				});
-
-				describe('delegate asset', () => {
-					it('should be an object', () => {
-						(delegateTransaction.asset).should.have.property('delegate').and.be.type('object');
-					});
-
-					it('should have the provided username as a string', () => {
-						(delegateTransaction.asset.delegate).should.have.property('username').and.be.type('string').and.equal(username);
-					});
+				it('should have the provided username as a string', () => {
+					(delegateTransaction.asset.delegate).should.have.property('username').and.be.type('string').and.equal(username);
 				});
 			});
 		});
+	});
 
-		describe('with second secret', () => {
-			beforeEach(() => {
-				delegateTransaction = createDelegate(secret, username, secondSecret);
+	describe('with second secret', () => {
+		beforeEach(() => {
+			delegateTransaction = createDelegate(secret, username, secondSecret);
+		});
+
+		it('should create a delegate transaction with a second secret', () => {
+			const delegateTransactionWithoutSecondSecret = createDelegate(secret, username);
+			(delegateTransaction).should.be.ok();
+			(delegateTransaction).should.not.be.equal(delegateTransactionWithoutSecondSecret);
+		});
+
+		describe('returned delegate transaction', () => {
+			it('should have second signature hex string', () => {
+				(delegateTransaction).should.have.property('signSignature').and.be.hexString();
 			});
 
-			it('should create a delegate transaction with a second secret', () => {
-				const delegateTransactionWithoutSecondSecret = createDelegate(secret, username);
-				(delegateTransaction).should.be.ok();
-				(delegateTransaction).should.not.be.equal(delegateTransactionWithoutSecondSecret);
+			it('should be second signed correctly', () => {
+				const result = cryptoModule.verifyTransaction(delegateTransaction, secondPublicKey);
+				(result).should.be.ok();
 			});
 
-			describe('returned delegate transaction', () => {
-				it('should have second signature hex string', () => {
-					(delegateTransaction).should.have.property('signSignature').and.be.hexString();
-				});
-
-				it('should be second signed correctly', () => {
-					const result = cryptoModule.verifyTransaction(delegateTransaction, secondPublicKey);
-					(result).should.be.ok();
-				});
-
-				it('should not be second signed correctly if modified', () => {
-					delegateTransaction.amount = 100;
-					const result = cryptoModule.verifyTransaction(delegateTransaction, secondPublicKey);
-					(result).should.not.be.ok();
-				});
+			it('should not be second signed correctly if modified', () => {
+				delegateTransaction.amount = 100;
+				const result = cryptoModule.verifyTransaction(delegateTransaction, secondPublicKey);
+				(result).should.not.be.ok();
 			});
 		});
 	});
